Fixed listener test at line 54 in fake/tests.rs
The original listener test was asserting only the first argument of the .schedule_upcall which is not the sample.
Modifying the function to set sample as the first argument conflicts with the assessments in /peripherals/adc/src/test.rs
The test was modified to check for the third argument which was the where the sample should be checked for and now the test asserts the value correctly.

Signed-off-by: LithTheFox thelittrion@gmail.com